Assistant_support_LangGraph/Project/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain
python-dotenv
python-multipart
requests